+++ test/VaultSharp.Samples/Backends/Auth/AliCloudAuthBackendSamples.cs
@@ -1,0 +1,10 @@
+﻿namespace VaultSharp.Samples
+{
+    partial class Program
+    {
+        private static void RunAliCloudAuthMethodSamples()
+        {
+            // raja todo.
+        }
+    }
+}